refactor(contracts): use @contract-types alias
Replaced all imports from `@typechain` with `@contract-types` to improve
semantics and prepare for future refactoring. This change simplifies the
codebase and makes it easier to understand the purpose of the imported
modules.

The `package.json` file was also updated to reflect these changes,
ensuring that all paths and exports are correctly configured.

Signed-off-by: Miguel_LZPF <miguel.carpena@io.builders>